Create onComposerShown.ts
